LP5521 LED: Use a work queue for button backlight fading
The fading was implemented in a blocking way so that a change took around 200ms. This was
causing a noticeable input lag in the GUI because the buttons are dimmed when the user is
inactive and parts of the Android system wait for the liblights/LightsService call to
finish.

The time to change the brightness can be easily measured and verified with:

    time echo "255" > /sys/class/leds/button-backlight/brightness

Conflicts:
	drivers/leds/leds-lp5521_htc.c