LP5521 LED: Add slow blink mode for button-backlight LED
Add a new "slow_blink" sysfs entry for each LED to control the slow blinking
behavior. At the moment only the button-backlight LED is supported. For each
value between 1-255 a backlight program is created that fades in to the passed
value, waits 1s, fades out, wait 3s and repeats from the start. Writing a 0
will turn the button backlight off immediately.

Only 32 instructions can be used for the backlight program. It was not tested if
there are sleeping instructions > 1s and why the extra 0x00 data is needed after
each instruction.

Patch Set 4: Add some debugging prints and fix resetting backlight blink mode

For some modes the incorrect register was written when backlight_mode was set to
0. Also wait longer after writing to the ENABLE and OP_MODE registers which should
fix button backlight not turning on in some cases.

Also spaces -> tabs.

Change-Id: Ib1b2f0db55de466595cf16c119594b4fbc6140fb